Refactor. change naming utils=>util and brush up exporort expression
