+++ src/util/multerErrorHandler.js
@@ -1,7 +1,7 @@
 const CreateError = require("http-errors");
 const { ONLY_MP4_MOV_ALLOWED, VIDEO_TOO_LARGE } = require("../constants/error");
 
-const multerErrorHandler = (error, req, res, next) => {
+exports.multerErrorHandler = (error, req, res, next) => {
   if (error.message === "Only mp4 and mov are allowed") {
     throw CreateError(415, ONLY_MP4_MOV_ALLOWED);
   }
@@ -12,5 +12,3 @@ const multerErrorHandler = (error, req, res, next) => {
 
   next();
 };
-
-module.exports = multerErrorHandler;
